Add jekyll feed to dependencies.
I want to support RSS integration because I think
integration with aggregators is important. Content
aggregation that requires people to go to my site
wastes the time of people who have better things to
do. It also makes it harder for aggregation services
to notify people of updates to my content.