Add link to head so RSS readers can detect changes.
Not all users are going to use RSS feeds, but making it
so that those who do have their browser setup to let
them know about RSS existence on a page should lower
the burden for people who want to follow me to do so.

Tested that RSS reader extension detects RSS presence.