Add meta for follow.it
This adds a meta tag requested by follow.it which allows me
to claim ownership of the subscription service.

I also updated the comments in the header to better reflect
the intended organization so that working with the head in
the future is easier to do.

These changes haven't been tested yet.