Add support for subscribing to the site via email.
RSS has some really terrible ergonomics when we don't
outsource its fulfillment. I wrote a little about this
problem and a long term solution for it here:
https://twitter.com/jcolechanged/status/1363205840207896576

However, in the short term we don't have a good open
ecosystem for subscription fufillment. So if I want to
save people who use my site time I need to figure out a
place that can do fufillment for them.

I went with follow.it because I've seen it being used on
another blog I follow and because they have a good name.

From this point on the site is also going to be using
colors to help facillitate effective communication. Orange
will be for notifications, because orange is the color of
the default RSS icon which is supposed to be represenative
of the idea of notifications being sent out over radio waves.