Fix footnotes failing to render quines.
One of things I consider important is being upfront about
how the visualizations I'm creating are being produced.
I want my visualizations to be backed by data that people
can trace the lineage of and I want my visualizations to
be created such that people can see that I'm not doing
anything underhanded to make the data tell a different
story than ought it to be telling. More then that I'm
trying to reduce the feeling of magic that visualizations
can produce by pulling back the curtain, because magic
and explanation are at odds with each other.

One of the ways I'm accomplishing this is through the
use of a quine tool. The quine tool loads the code I
write which does visualization and places it on the page.
This mechanism leads to quine construction happening after
page render, because the same code that does the visualization
also does the quine reproduction.

Popfoots previously failed to integrate with this, because it
ran before quines were created. This has been addressed in
this commit by moving the popfoots setup to begin only once the
page has been loaded and other document ready functions have
finished running.

I've verified on the page that the quine is now rendered. The
next step in improving popfoots will be improving the visualization
and supporting jumping to the footnote in addition to rendering the
footnote inline as well as allowing for jumpbacks.