Give popnotes some room to breathe and more congruent coloring.
Popnotes were showing up in gray. This made the notes jarring.
This fixes that issue by making them colored similar to the rest
of the page. It also refactors them a bit; they are now decidedly
not mobile friendly and very much take advantage of a large window.

Verified that the popup notes are now much more pleasing.